Issue #SC-1168 fix: merging keycloak util code to user-org util and cleaning unused code
