--- user-org-keycloak-util/src/main/java/org/sunbird/sso/model/User.java
@@ -1,100 +1,0 @@
-/** */
-package org.sunbird.sso.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.Map;
-
-/** @author manzarul */
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class User {
-
-  private String firstName;
-  private String lastName;
-  private String email;
-  private String phone;
-  private String countryCode;
-  private boolean phoneVerified;
-  private boolean emailVerified;
-  private String userName;
-  private String id;
-  private static ObjectMapper mapper = new ObjectMapper();
-
-  public User() {}
-
-  public String getFirstName() {
-    return firstName;
-  }
-
-  public void setFirstName(String firstName) {
-    this.firstName = firstName;
-  }
-
-  public String getLastName() {
-    return lastName;
-  }
-
-  public void setLastName(String lastName) {
-    this.lastName = lastName;
-  }
-
-  public String getEmail() {
-    return email;
-  }
-
-  public void setEmail(String email) {
-    this.email = email;
-  }
-
-  public String getPhone() {
-    return phone;
-  }
-
-  public void setPhone(String phone) {
-    this.phone = phone;
-  }
-
-  public String getCountryCode() {
-    return countryCode;
-  }
-
-  public void setCountryCode(String countryCode) {
-    this.countryCode = countryCode;
-  }
-
-  public boolean isPhoneVerified() {
-    return phoneVerified;
-  }
-
-  public void setPhoneVerified(boolean phoneVerified) {
-    this.phoneVerified = phoneVerified;
-  }
-
-  public boolean isEmailVerified() {
-    return emailVerified;
-  }
-
-  public void setEmailVerified(boolean emailVerified) {
-    this.emailVerified = emailVerified;
-  }
-
-  public String getUserName() {
-    return userName;
-  }
-
-  public void setUserName(String userName) {
-    this.userName = userName;
-  }
-
-  public String getId() {
-    return id;
-  }
-
-  public void setId(String id) {
-    this.id = id;
-  }
-
-  public static User initializeUserFromMap(Map<String, Object> userReqMap) {
-    return mapper.convertValue(userReqMap, User.class);
-  }
-}
