--- user-org-keycloak-util/src/test/resources/sso.properties
@@ -1,7 +1,0 @@
-sunbird_sso_url=http://localhost:8080/auth/
-sunbird_sso_realm=master
-sunbird_sso_username=manzarul
-sunbird_sso_password=password
-sunbird_sso_client_id=admin-cli
-sunbird_sso_pool_size=10
-sso.enabled=true
