+++ user-org-util/src/test/resources/sso.properties
